extracting decision about data cell info to another separated class to be used anywhere
