+++ core/src/main/java/at/specure/info/cell/ActiveDataCellInfoExtractor.kt
@@ -1,0 +1,8 @@
+package at.specure.info.cell
+
+import android.telephony.CellInfo
+
+interface ActiveDataCellInfoExtractor {
+
+    fun extractActiveCellInfo(cellInfo: MutableList<CellInfo>): ActiveDataCellInfo
+}
